now setcolour carries date
