+++ src/headers/setColour.h
@@ -1,7 +1,12 @@
 #ifndef SETCOLOUR_H
 #define SETCOLOUR_H
 
+#include <iostream>
+
+using namespace std;
+
 void setColour ( int colourCode );
 void resetColour();
+string generateDate ();
 
 #endif
